Move config and cli components to separate packages
