+++ csb/src/main/scala/edu/msstate/dasi/csb/Benchmark.scala
@@ -1,5 +1,7 @@
 package edu.msstate.dasi.csb
 
+import edu.msstate.dasi.csb.cli.OptionParser
+import edu.msstate.dasi.csb.config.{ComponentFactory, Config}
 import edu.msstate.dasi.csb.data.distributions.DataDistributions
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
 import edu.msstate.dasi.csb.data.seed.DataParser
