+++ csb/src/main/scala/edu/msstate/dasi/csb/config/ComponentFactory.scala
@@ -1,7 +1,7 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.config
 
-import edu.msstate.dasi.csb.persistence.{GraphPersistence, Neo4jPersistence, SparkPersistence}
 import edu.msstate.dasi.csb.data.synth.{GraphSynth, ParallelBa, StochasticKronecker}
+import edu.msstate.dasi.csb.persistence.{GraphPersistence, Neo4jPersistence, SparkPersistence}
 import edu.msstate.dasi.csb.veracity._
 import edu.msstate.dasi.csb.workload.{Neo4jWorkload, SparkWorkload, Workload}
